+++ controllers/borrowerController.js
@@ -1,13 +1,14 @@
+const AppError = require("../utils/appError");
 
 // Get All Borrowers
 exports.getAllBorrowers = (req, res, next) => {
-    // Construct the base SQL query
-    let sql = 'SELECT * FROM borrowers ';
     const db = req.app.locals.db;
+
+    let sql = 'SELECT * FROM borrowers ';
 
     db.query(sql, (err, results) => {
         if (err) {
-            return next(err); // Pass the error to the error-handling middleware
+            return next(new AppError(err.message, 400));
         }
         res.status(200).json({ status: "Success", length: results.length, results });
     });
@@ -16,17 +17,22 @@ exports.getAllBorrowers = (req, res, next) => {
 
 // Get Borrower by ID
 exports.getBorrower = (req, res, next) => {
-    const borrowerId  = req.params.id;
     const db = req.app.locals.db;
+
+    const borrowerId = Number(req.params.id);
+    // make sure passed id is a number
+    if (typeof borrowerId !== 'number' || isNaN(borrowerId)) {
+        return next(new AppError('Borrower\'s ID should be a numeric value', 400));
+    }
 
     let sql = 'SELECT * FROM borrowers WHERE id = ?';
 
     db.query(sql, [borrowerId], (err, results) => {
         if (err) {
-            return next(err); // Pass the error to the error-handling middleware
+            return next(new AppError(err.message, 400));
         }
         if (results.length === 0) {
-            return res.status(404).json({ error: "Borrower not found" });
+            return res.status(404).json({ status: "Failed", message: "Borrower not found" });
         }
         res.status(200).json({ status: "Success", result: results[0] });
     });
@@ -40,11 +46,28 @@ exports.createBorrower = (req, res, next) => {
 
     // Check if all required fields are present
     if (!name || !email) {
-        return res.status(400).json({ error: "All fields are required" });
+        return next(new AppError('Name and Email fields are all required', 400));
     }
+
+    // Validate email format
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if (!emailRegex.test(email)) {
+        return next(new AppError('Invalid email format', 400));
+    }
+
     db.query('INSERT INTO borrowers (name, email, reg_date) VALUES (?, ?, NOW())', [name, email], (err, result) => {
-        if (err) throw err;
-        res.status(201).json({ status: "Success", result });
+        if (err) return next(new AppError(err.message, 500));
+        // Fetch the newly inserted borrower from the database
+        db.query('SELECT * FROM borrowers WHERE id = ?', [result.insertId], (fetchErr, fetchResult) => {
+            if (fetchErr) return next(new AppError(fetchErr.message, 500));
+
+            const insertedBorrower = fetchResult[0]; // Assuming fetchResult is an array of rows
+            res.status(201).json({
+                status: "Success",
+                message: "Borrower is successfully inserted into database",
+                borrower: insertedBorrower
+            });
+        });
     });
 };
 
@@ -52,22 +75,32 @@ exports.createBorrower = (req, res, next) => {
 exports.updateBorrower = (req, res, next) => {
     const db = req.app.locals.db;
 
-    const borrowerId = req.params.id;
+    const borrowerId = Number(req.params.id);
+    // make sure passed id is a number
+    if (typeof borrowerId !== 'number' || isNaN(borrowerId)) {
+        return next(new AppError('Borrower\'s ID should be a numeric value', 400));
+    }
+
     const { name, email } = req.body;
 
     // Check if any update fields are provided
     if (!name && !email) {
-        return res.status(400).json({ error: "At least one field is required for update" });
+        return next(new AppError("At least one field is required for update", 400));
+    }
+    // Validate email format
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if (email && !emailRegex.test(email)) {
+        return next(new AppError('Invalid email format', 400));
     }
 
     // Construct the SQL query for updating the borrower
     let sql = 'UPDATE borrowers SET ';
     const updateValues = [];
-    if (title) {
+    if (name) {
         sql += 'name = ?, ';
         updateValues.push(name);
     }
-    if (author) {
+    if (email) {
         sql += 'email = ?, ';
         updateValues.push(email);
     }
@@ -81,20 +114,32 @@ exports.updateBorrower = (req, res, next) => {
     // Execute the SQL query
     db.query(sql, updateValues, (err, result) => {
         if (err) {
-            return next(err); // Pass the error to the error-handling middleware
+            return next(new AppError(err.message, 500));
         }
         if (result.affectedRows === 0) {
-            return res.status(404).json({ error: "Borrower not found" });
+            return res.status(404).json({ status: "Failed", message: "Borrower not found" })
         }
-        res.json({ status: "Success", message: "Borrower updated successfully" });
+        // Fetch the updated borrower from the database
+        db.query('SELECT * FROM borrowers WHERE id = ?', [borrowerId], (fetchErr, fetchResult) => {
+            if (fetchErr) {
+                return next(new AppError(fetchErr.message, 500));
+            }
+
+            const updatedBorrower = fetchResult[0]; // Assuming fetchResult is an array of rows
+            res.status(200).json({ status: "Success", message: "Borrower updated successfully", borrower: updatedBorrower });
+        });
     });
 };
 
 // Delete a Borrower
 exports.deleteBorrower = (req, res, next) => {
-    const borrowerId = req.params.id;
     const db = req.app.locals.db;
 
+    const borrowerId = Number(req.params.id);
+    // make sure passed id is a number
+    if (typeof borrowerId !== 'number' || isNaN(borrowerId)) {
+        return next(new AppError('Borrower\'s ID should be a numeric value', 400));
+    }
 
     // Construct the SQL query for deleting the borrower by its ID
     const sql = 'DELETE FROM borrowers WHERE id = ?';
@@ -102,11 +147,16 @@ exports.deleteBorrower = (req, res, next) => {
     // Execute the SQL query
     db.query(sql, [borrowerId], (err, result) => {
         if (err) {
-            return next(err); // Pass the error to the error-handling middleware
+            // Check if the error message indicates a foreign key constraint failure
+            if (err.message.includes('foreign key constraint fails')) {
+                return next(new AppError('Cannot delete borrower: related records exist in borrowing processes', 400));
+            }
+            // For other errors, pass the error to the error-handling middleware
+            return next(new AppError(err.message, 500));
         }
         if (result.affectedRows === 0) {
-            return res.status(404).json({ error: "Borrower not found" });
+            return res.status(404).json({ status: "fail", message: "Borrower not found" });
         }
-        res.json({ status: "Success", message: "Borrower deleted successfully" });
+        res.status(204).json({ status: "success", message: "Borrower deleted successfully" });
     });
 };
